refactor: replace index.ts to the root of the redux folder
