+++ kanban-app/src/app/core/services/notification.service.ts
@@ -1,8 +1,6 @@
 import { Injectable } from '@angular/core';
 import { Store } from '@ngrx/store';
-import { BehaviorSubject, Subject } from 'rxjs';
 import { addNotification } from 'src/app/redux/actions/notification.actions';
-import { State } from 'src/app/redux/reducers';
 import { selectNotificationMessage } from 'src/app/redux/selectors/notification.selectors';
 
 @Injectable({
@@ -11,7 +9,7 @@ import { selectNotificationMessage } from 'src/app/redux/selectors/notification.
 export class NotificationService {
   notification$ = this.store.select(selectNotificationMessage);
 
-  constructor(private store: Store<State>) {}
+  constructor(private store: Store) {}
 
   setNotification(notification: string) {
     this.store.dispatch(addNotification({ message: notification }));
