+++ kanban-app/src/app/redux/index.ts
@@ -1,14 +1,11 @@
 import {
-  ActionReducer,
   ActionReducerMap,
-  createFeatureSelector,
-  createSelector,
   MetaReducer,
 } from '@ngrx/store';
-import { environment } from '../../../environments/environment';
-import { StateConfirm } from './confirm.reducer';
-import { StateNotification } from './notification.reducer';
-import { StateUser } from './user.reducer';
+import { environment } from 'src/environments/environment';
+import { StateConfirm } from './reducers/confirm.reducer';
+import { StateNotification } from './reducers/notification.reducer';
+import { StateUser } from './reducers/user.reducer';
 
 export interface State {
   notification?: StateNotification;
